[REFACTOR] 카테고리 모달 내용 파일 분리
